change all folders
